fix -Wsign-conversion warnings (#1106)
Adds to_s32() overloads for const char*/char* (previously only to_u32()
parsed strings, so a signed env var had to round-trip through unsigned),
promotes test_multiprocess child indices to size_t with a bounds check at
the Windows argv parse boundary, and passes unsigned queue flags at two
create_command_bundle() call sites.

+++ stress_tests/test_multiprocess/src/main.cpp
@@ -20,7 +20,7 @@
 #include <boost/interprocess/shared_memory_object.hpp>
 
 // Forward declaration - definition lives in test_multiprocess.cpp.
-int child_work(int child_index, ChildResults *shm,
+int child_work(size_t child_index, ChildResults *shm,
                const std::string &worker_name);
 
 // Returns the child exit code when running as a Windows child worker,
@@ -46,7 +46,10 @@ static int maybe_run_as_child_worker(int argc, char **argv) {
                                        bipc::read_write);
     bipc::mapped_region region(shm_obj, bipc::read_write);
     ChildResults *shm = static_cast<ChildResults *>(region.get_address());
-    return child_work(child_index, shm, std::string(worker_fn));
+    // child_index is validated non-negative above; child_work() bounds-checks
+    // it against the shared-memory capacity.
+    return child_work(static_cast<size_t>(child_index), shm,
+                      std::string(worker_fn));
   } catch (...) {
     return 1;
   }
